requirements.txt
# web / framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# pydantic v2 + settings
pydantic==2.11.9
pydantic-settings==2.10.1

# DB / async
SQLAlchemy==2.0.43
asyncpg==0.30.0

# cache / pubsub
redis==6.4.0

# utilities / dev
python-dotenv==1.1.1
httpx==0.28.1
typing-extensions==4.15.0
pytest==8.4.2
